master_ expanding the coins tracked. Added more currency conversion and formatted data to only include data needed. Only creating one file per day now

+++ BtcTracker/src/tickers/CryptoWatchTicker.kt
@@ -9,6 +9,6 @@ data class Result(
 data class PriceHolder(val price: Price)
 
 data class Price(
-        var last: Float? = null,
-        var high: Float? = null,
-        var low: Float? = null)
+        var last: Double? = null,
+        var high: Double? = null,
+        var low: Double? = null)
